Refactor to support semi-lazy loading
There should always be a password if we actually care about encryption.
If If there isn't a password, then the build should only fail if
encryption/decryption is actually requested, e.g. if the credentials are
only used to publish the build should work otherwise even if the
password isn't available.