Added the following @Options (project properties):   --env (credentialsEnv)   --loc (credentialsLocation - *preexisting)   --pass (credentialsPassphrase - *preexisting)
Added --hashed=<boolean default = true> to AddCredentialsTask to indicate that the key should be hashed (more secure). use --hashed=[no|off|false|blah] to disable

ShowCredentialsTask: Look up a single key without having to define a custom task in your personal build.

Crypto: The magic happens here.

Keys are now hashed with the input passphrase. This eliminates the leakage of security by naming the file after the md5 of the passphrase.

Environments allow you to separate properties without relying on a different password as the means to create a different properties file. The encrypted properties filename is a hash of the environment and the passphrase together. (It isn't 100% perfect but it does hinder rainbow table attacks - provided the defaults are not chosen for real secrets)

Changed:
 Salts and ivs are now generated for every value encrypted. No more reusing the same salt/predictable iv. Added the ability to define @Options for AddCredentialsTask, RemoveCredentialsTask, and ShowCredentialsTask. Any @Option will override any defined project level property. Drastically changed the Encryption class to the point that I felt it should be changed altogether.

 Kept:
  Strived for 100% backwards compatibility. Added a number of tests to ensure as much.

 Removed:
  The naming of the gradle encrypted properties has changed slightly. It is no longer tied directly to the MD5HASH of the passphrase. It is now a hash of the passphrase and the environment (--env/credentialsEnv) values.